Extract GearRatioChart into separate file
Move the inline GearRatioChart implementation out of TuneForm/TuneSettingsPanel into a dedicated component file (client/src/components/tune/GearRatioChart.tsx). Update TuneForm and TuneSettingsPanel to import and use the new component. This removes duplicated chart code, centralizes chart logic (including tire circumference/back-calculation, gridlines, and rendering), and introduces support for both topSpeedKph and topSpeedMph props while preserving defaults (maxRpm = 8000). The change is mostly a refactor to reduce duplication and improve reuse.
