+++ client/src/components/TuneForm.tsx
@@ -4,6 +4,7 @@ import { useQuery } from "@tanstack/react-query";
 import type { TuneSettings } from "../data/tune-catalog";
 import type { TuneCategory } from "@shared/types";
 import { client } from "../lib/rpc";
+import { GearRatioChart } from "./tune/GearRatioChart";
 
 // ── Hook ─────────────────────────────────────────────────────────────────────
 
@@ -501,215 +502,6 @@ export function UserTuneCard({
 				</div>
 			)}
 		</div>
-	);
-}
-
-// ── GearRatioChart ────────────────────────────────────────────────────────────
-
-function GearRatioChart({
-	ratios,
-	finalDrive,
-	topSpeedMph,
-	maxRpm = 8000,
-}: {
-	ratios: number[];
-	finalDrive: number;
-	topSpeedMph?: number;
-	maxRpm?: number;
-}) {
-	if (!ratios.length) return null;
-
-	const topGearRatio = ratios[ratios.length - 1];
-	// Back-calculate tire circumference from stock top speed if available, else use typical 2.0m
-	const CIRC =
-		topSpeedMph && topGearRatio
-			? (topSpeedMph * 1.60934 * topGearRatio * finalDrive) /
-				(maxRpm / 60) /
-				3.6
-			: 2.0;
-	const toKph = (rpm: number, ratio: number) =>
-		(rpm / 60 / (ratio * finalDrive)) * CIRC * 3.6;
-
-	const maxSpeed =
-		Math.ceil(toKph(maxRpm, ratios[ratios.length - 1]) / 50) * 50;
-
-	const W = 280,
-		H = 120;
-	const pad = { top: 18, right: 16, bottom: 24, left: 32 };
-	const cW = W - pad.left - pad.right;
-	const cH = H - pad.top - pad.bottom;
-
-	const sx = (v: number) => Math.min((v / maxSpeed) * cW, cW);
-	const sy = (rpm: number) => cH - (rpm / maxRpm) * cH;
-
-	// RPM gridlines at even intervals
-	const rpmStep = maxRpm <= 8000 ? 2000 : maxRpm <= 12000 ? 3000 : 4000;
-	const rpmGrids = Array.from(
-		{ length: Math.floor(maxRpm / rpmStep) },
-		(_, i) => (i + 1) * rpmStep,
-	);
-	const speedGridCount = 4;
-	const speedGrids = Array.from({ length: speedGridCount + 1 }, (_, i) =>
-		Math.round((maxSpeed / speedGridCount) * i),
-	);
-
-	// Redline at maxRpm
-	const redlineY = pad.top + sy(maxRpm);
-
-	return (
-		<svg
-			width="100%"
-			viewBox={`0 0 ${W} ${H}`}
-			style={{ display: "block", maxWidth: "280px" }}
-		>
-			<defs>
-				<clipPath id="gchart">
-					<rect x={pad.left} y={pad.top} width={cW} height={cH} />
-				</clipPath>
-			</defs>
-
-			{/* Background */}
-			<rect
-				x={pad.left}
-				y={pad.top}
-				width={cW}
-				height={cH}
-				fill="currentColor"
-				fillOpacity="0.03"
-			/>
-
-			{/* RPM gridlines */}
-			{rpmGrids.map((rpm) => (
-				<g key={rpm}>
-					<line
-						x1={pad.left}
-						y1={pad.top + sy(rpm)}
-						x2={pad.left + cW}
-						y2={pad.top + sy(rpm)}
-						stroke="currentColor"
-						strokeOpacity="0.1"
-						strokeWidth="1"
-					/>
-					<text
-						x={pad.left - 4}
-						y={pad.top + sy(rpm) + 3}
-						textAnchor="end"
-						fontSize="7"
-						fill="currentColor"
-						fillOpacity="0.45"
-					>
-						{rpm / 1000}
-					</text>
-				</g>
-			))}
-
-			{/* Speed gridlines */}
-			{speedGrids.map((spd) => (
-				<g key={spd}>
-					<line
-						x1={pad.left + sx(spd)}
-						y1={pad.top}
-						x2={pad.left + sx(spd)}
-						y2={pad.top + cH}
-						stroke="currentColor"
-						strokeOpacity="0.1"
-						strokeWidth="1"
-					/>
-					<text
-						x={pad.left + sx(spd)}
-						y={pad.top + cH + 10}
-						textAnchor="middle"
-						fontSize="7"
-						fill="currentColor"
-						fillOpacity="0.45"
-					>
-						{spd}
-					</text>
-				</g>
-			))}
-
-			{/* Axis labels */}
-			<text
-				x={pad.left + cW}
-				y={pad.top + cH + 20}
-				textAnchor="end"
-				fontSize="7"
-				fill="currentColor"
-				fillOpacity="0.35"
-			>
-				KM/H
-			</text>
-			<text
-				x={pad.left - 4}
-				y={pad.top - 6}
-				textAnchor="end"
-				fontSize="7"
-				fill="currentColor"
-				fillOpacity="0.35"
-			>
-				RPM ×1000
-			</text>
-
-			{/* Gear lines — white, slightly transparent, each starts from prev gear's shift point */}
-			{ratios.map((ratio, i) => {
-				const startKph = i === 0 ? 0 : toKph(maxRpm, ratios[i - 1]);
-				const startRpm =
-					i === 0 ? 0 : (((startKph / 3.6) * (ratio * finalDrive)) / CIRC) * 60;
-				const steps = 60;
-				const pts = Array.from({ length: steps }, (_, j) => {
-					const rpm = startRpm + (j / (steps - 1)) * (maxRpm - startRpm);
-					return `${pad.left + sx(toKph(rpm, ratio))},${pad.top + sy(rpm)}`;
-				}).join(" ");
-				const labelX = pad.left + sx(toKph(maxRpm, ratio));
-				return (
-					<g key={i}>
-						<polyline
-							points={pts}
-							fill="none"
-							stroke="white"
-							strokeWidth="1.5"
-							strokeOpacity="0.7"
-							clipPath="url(#gchart)"
-						/>
-						<text
-							x={labelX + 2}
-							y={pad.top + sy(maxRpm) - 3}
-							textAnchor="middle"
-							fontSize="7"
-							fill="white"
-							fillOpacity="0.6"
-							fontWeight="600"
-						>
-							{i + 1}
-						</text>
-					</g>
-				);
-			})}
-
-			{/* Redline */}
-			<line
-				x1={pad.left}
-				y1={redlineY}
-				x2={pad.left + cW}
-				y2={redlineY}
-				stroke="#ef4444"
-				strokeWidth="1"
-				strokeOpacity="0.8"
-				strokeDasharray="3 2"
-			/>
-
-			{/* Chart border */}
-			<rect
-				x={pad.left}
-				y={pad.top}
-				width={cW}
-				height={cH}
-				fill="none"
-				stroke="currentColor"
-				strokeOpacity="0.15"
-				strokeWidth="1"
-			/>
-		</svg>
 	);
 }
